+++ packages/squared-android-template/content/template/.sdl-pg.lua
@@ -3,9 +3,9 @@
 return {
     format = 3,
     generator = "squared-pg",
-    generator_version = "0.6.0-dev.5",
+    generator_version = "0.6.0-dev.6",
     name = "{{PROJECT_NAME}}",
     package = "{{PACKAGE_NAME}}",
     version = "{{BASE_VERSION}}",
-    template = "dev.squarednetizen.template.android-sdl2-lua@0.6.0-dev.14"
+    template = "dev.squarednetizen.template.android-sdl2-lua@0.6.0-dev.15"
 }