+++ packages/squared-android-template/content/template/project.lua
@@ -7,8 +7,8 @@ return {
     package = "{{PACKAGE_NAME}}",
     version = "{{BASE_VERSION}}",
     profile = "android-sdl2-lua",
-    template = "dev.squarednetizen.template.android-sdl2-lua@0.6.0-dev.14",
-    generator_version = "0.6.0-dev.5",
+    template = "dev.squarednetizen.template.android-sdl2-lua@0.6.0-dev.15",
+    generator_version = "0.6.0-dev.6",
     plugin_api_version = 1,
     offline_first = true
 }